Move core code to core package
