+++ src/main/scala/com/chrisomeara/pillar/core/ConfigurationException.scala
@@ -1,3 +1,3 @@
-package com.chrisomeara.pillar
+package com.chrisomeara.pillar.core
 
 class ConfigurationException(message: String) extends RuntimeException(message)
