+++ src/main/scala/com/chrisomeara/pillar/core/InvalidMigrationException.scala
@@ -1,3 +1,3 @@
-package com.chrisomeara.pillar
+package com.chrisomeara.pillar.core
 
 class InvalidMigrationException(val errors: Map[String,String]) extends RuntimeException
