+++ src/main/scala/com/chrisomeara/pillar/core/ReportingMigration.scala
@@ -1,6 +1,7 @@
-package com.chrisomeara.pillar
+package com.chrisomeara.pillar.core
 
 import java.util.Date
+
 import com.datastax.driver.core.Session
 
 class ReportingMigration(reporter: Reporter, wrapped: Migration) extends Migration {
