+++ src/test/scala/com/chrisomeara/pillar/core/ReportingMigrationSpec.scala
@@ -1,11 +1,10 @@
-package com.chrisomeara.pillar
+package com.chrisomeara.pillar.core
 
+import com.datastax.driver.core.Session
+import org.mockito.Mockito._
 import org.scalatest.FunSpec
 import org.scalatest.matchers.ShouldMatchers
 import org.scalatest.mock.MockitoSugar
-import org.mockito.Mockito._
-import com.datastax.driver.core.Session
-import java.util.Date
 
 class ReportingMigrationSpec extends FunSpec with ShouldMatchers with MockitoSugar {
   val reporter = mock[Reporter]
