+++ src/test/scala/com/chrisomeara/pillar/core/ReportingMigratorSpec.scala
@@ -1,4 +1,4 @@
-package com.chrisomeara.pillar
+package com.chrisomeara.pillar.core
 
 import com.datastax.driver.core.Session
 import org.mockito.Mockito._
